api_framework/requirements.txt
pytest==7.4.3
requests==2.31.0
pytest-html==4.1.1
pytest-xdist==3.3.1
allure-pytest==2.13.2
python-dotenv==1.0.0
pydantic>=2.7.0
pydantic-settings>=2.4.0
jsonschema==4.20.0
faker==20.1.0
